% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_sub2.R
\name{makeLabels}
\alias{makeLabels}
\title{Make labels for ROC curve}
\usage{
makeLabels(x, no = 1)
}
\arguments{
\item{x}{A list As a result of Epi::ROC()}

\item{no}{Integer}
}
\value{
A data.frame
}
\description{
Make labels for ROC curve
}
\examples{
require(Epi)
x=ROC(form=am~wt,data=mtcars,plot="ROC")
makeLabels(x)
}
