% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_tree.R
\name{makePPTList_surTree}
\alias{makePPTList_surTree}
\title{Make Powerpoint List for survival tree analysis}
\usage{
makePPTList_surTree(fitstr)
}
\arguments{
\item{fitstr}{string A string or R code}
}
\value{
A data.frame
}
\description{
Make Powerpoint List for survival tree analysis
}
\examples{
require(party)
data(GBSG2,package="TH.data")
result=makePPTList_surTree("ctree(Surv(time,cens)~.,data=GBSG2)")
}
