% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_tree.R
\name{makePPTList_tree}
\alias{makePPTList_tree}
\title{Make Powerpoint List for tree analysis}
\usage{
makePPTList_tree(fitstr)
}
\arguments{
\item{fitstr}{string A string or R code}
}
\value{
A data.frame
}
\description{
Make Powerpoint List for tree analysis
}
\examples{
require(party)
result=makePPTList_tree("ctree(Species~.,data=iris)")
}
