% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_sub2.R
\name{optimal_cutpoint}
\alias{optimal_cutpoint}
\title{Calculate optimal cutpoint}
\usage{
optimal_cutpoint(x)
}
\arguments{
\item{x}{A list As a result of Epi::ROC()}
}
\description{
Calculate optimal cutpoint
}
\examples{
require(Epi)
x=ROC(form=am~wt,data=mtcars,plot="ROC")
optimal_cutpoint(x)
}
