% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_sub2.R
\name{optimal_lr.eta}
\alias{optimal_lr.eta}
\title{Calculate optimal lr.eta}
\usage{
optimal_lr.eta(x)
}
\arguments{
\item{x}{A list As a result of Epi::ROC()}
}
\description{
Calculate optimal lr.eta
}
\examples{
require(Epi)
x=ROC(form=am~wt,data=mtcars,plot="ROC")
optimal_lr.eta(x)
}
