% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_sub2.R
\name{plot_ROC}
\alias{plot_ROC}
\title{Draw ROC curves}
\usage{
plot_ROC(
  x,
  show.line = FALSE,
  show.points = TRUE,
  show.lr.eta = TRUE,
  show.sens = TRUE,
  show.AUC = TRUE
)
}
\arguments{
\item{x}{A list of result(s) of Epi::ROC()}

\item{show.line}{logical}

\item{show.points}{logical}

\item{show.lr.eta}{logical}

\item{show.sens}{logical}

\item{show.AUC}{logical}
}
\value{
A ggplot
}
\description{
Draw ROC curves
}
\examples{
require(moonBook)
require(Epi)
a=ROC(form=male~height,data=radial,plot="")
plot_ROC(a)
b=ROC(form=male~weight,data=radial,plot="")
plot_ROC(list(a,b))
c=ROC(form=male~weight+height,data=radial,plot="")
plot_ROC(list(a,b,c),show.lr.eta=FALSE,show.sens=FALSE)
}
