% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_sub2.R
\name{ROC2roc}
\alias{ROC2roc}
\title{Convert A list from ROC() to roc object}
\usage{
ROC2roc(x)
}
\arguments{
\item{x}{A list As a result of Epi::ROC()}
}
\description{
Convert A list from ROC() to roc object
}
