% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTreeModule.R
\name{sTreeModule}
\alias{sTreeModule}
\title{Server function of sTree shiny module}
\usage{
sTreeModule(input, output, session, dataname, df, preprocessing, PPTdata, lang)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{dataname}{A data name}

\item{df}{A data.frame}

\item{preprocessing}{A character string of R code}

\item{PPTdata}{PPTdata}

\item{lang}{The language to be displayed}
}
\description{
Server function of sTree shiny module
}
